backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ── Google APIs ─────────────────────────────────────────────────
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth==2.37.0

# ── Security & Config ──────────────────────────────────────────
itsdangerous==2.2.0
python-dotenv==1.0.1
pydantic-settings==2.7.1

# ── HTTP Client ─────────────────────────────────────────────────
httpx==0.28.1

# ── Testing ─────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.2
